+++ esp-idf-svc/src/eth.rs
@@ -1032,7 +1032,7 @@ impl<'d, T> EthDriver<'d, T> {
         Ok(())
     }
 
-    /// Enables or disables promiscuous mode for the Ethernet driver.
+    /// Enables or disables promiscuous mode for the [`EthDriver`].
     ///
     /// When promiscuous mode is enabled, the driver captures all Ethernet frames
     /// on the network, regardless of their destination MAC address. This is useful for